receiver split WORK IN PROGRESS
