+++ pkg/component/component.go
@@ -97,6 +97,7 @@ var (
 	Downsample      = source{component: component{name: "downsample"}}
 	Replicate       = source{component: component{name: "replicate"}}
 	QueryFrontend   = source{component: component{name: "query-frontend"}}
+	ReceiveRoute    = source{component: component{name: "receive-route"}}
 	Debug           = sourceStoreAPI{component: component{name: "debug"}}
 	Receive         = sourceStoreAPI{component: component{name: "receive"}}
 	Rule            = sourceStoreAPI{component: component{name: "rule"}}
@@ -104,4 +105,6 @@ var (
 	Store           = storeAPI{component: component{name: "store"}}
 	UnknownStoreAPI = storeAPI{component: component{name: "unknown-store-api"}}
 	Query           = storeAPI{component: component{name: "query"}}
+
 )
+
