+++ pkg/route/handler_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) The Thanos Authors.
 // Licensed under the Apache License 2.0.
 
-package receive
+package route
 
 import (
 	"bytes"
@@ -21,9 +21,9 @@ import (
 	"github.com/pkg/errors"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/storage"
+	terrors "github.com/prometheus/prometheus/tsdb/errors"
 	"google.golang.org/grpc"
 
-	"github.com/thanos-io/thanos/pkg/errutil"
 	"github.com/thanos-io/thanos/pkg/store/labelpb"
 	"github.com/thanos-io/thanos/pkg/store/storepb"
 	"github.com/thanos-io/thanos/pkg/store/storepb/prompb"
@@ -43,7 +43,7 @@ func TestCountCause(t *testing.T) {
 		},
 		{
 			name: "nil multierror",
-			err:  errutil.MultiError([]error{}),
+			err:  terrors.MultiError([]error{}),
 			f:    isConflict,
 			out:  0,
 		},
@@ -60,7 +60,7 @@ func TestCountCause(t *testing.T) {
 		},
 		{
 			name: "non-matching multierror",
-			err: errutil.MultiError([]error{
+			err: terrors.MultiError([]error{
 				errors.New("foo"),
 				errors.New("bar"),
 			}),
@@ -69,7 +69,7 @@ func TestCountCause(t *testing.T) {
 		},
 		{
 			name: "nested non-matching multierror",
-			err: errors.Wrap(errutil.MultiError([]error{
+			err: errors.Wrap(terrors.MultiError([]error{
 				errors.New("foo"),
 				errors.New("bar"),
 			}), "baz"),
@@ -78,9 +78,9 @@ func TestCountCause(t *testing.T) {
 		},
 		{
 			name: "deep nested non-matching multierror",
-			err: errors.Wrap(errutil.MultiError([]error{
+			err: errors.Wrap(terrors.MultiError([]error{
 				errors.New("foo"),
-				errutil.MultiError([]error{
+				terrors.MultiError([]error{
 					errors.New("bar"),
 					errors.New("qux"),
 				}),
@@ -90,7 +90,7 @@ func TestCountCause(t *testing.T) {
 		},
 		{
 			name: "matching multierror",
-			err: errutil.MultiError([]error{
+			err: terrors.MultiError([]error{
 				storage.ErrOutOfOrderSample,
 				errors.New("foo"),
 				errors.New("bar"),
@@ -100,7 +100,7 @@ func TestCountCause(t *testing.T) {
 		},
 		{
 			name: "matching multierror many",
-			err: errutil.MultiError([]error{
+			err: terrors.MultiError([]error{
 				storage.ErrOutOfOrderSample,
 				conflictErr,
 				errors.New(strconv.Itoa(http.StatusConflict)),
@@ -112,7 +112,7 @@ func TestCountCause(t *testing.T) {
 		},
 		{
 			name: "nested matching multierror",
-			err: errors.Wrap(errutil.MultiError([]error{
+			err: errors.Wrap(terrors.MultiError([]error{
 				storage.ErrOutOfOrderSample,
 				errors.New("foo"),
 				errors.New("bar"),
@@ -122,8 +122,8 @@ func TestCountCause(t *testing.T) {
 		},
 		{
 			name: "deep nested matching multierror",
-			err: errors.Wrap(errutil.MultiError([]error{
-				errutil.MultiError([]error{
+			err: errors.Wrap(terrors.MultiError([]error{
+				terrors.MultiError([]error{
 					errors.New("qux"),
 					errors.New(strconv.Itoa(http.StatusConflict)),
 				}),
